backend/pyproject.toml
[project]
name = "pet-time-tracker"
version = "0.1.0"
description = "FastAPI backend for pet time tracker"
authors = [{ name = "Pet Time Tracker" }]
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "sqlmodel",
  "sqlalchemy",
  "passlib[bcrypt]",
  "bcrypt>=4.1.2",
  "python-jose[cryptography]",
  "python-multipart",
  "pydantic-settings",
  "pillow",
  "email-validator",
]

[project.optional-dependencies]
dev = ["pytest", "httpx", "anyio"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]

[tool.pytest.ini_options]
pythonpath = ["app"]
testpaths = ["tests"]
